Moved emptyActivity and changed to correct naming
